delete 11 and rename 11a to 11 and fix solution
