--- samples/csharp_dotnetcore/11.qnamaker/BotServices.cs
@@ -1,38 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-using Microsoft.Bot.Builder.AI.QnA;
-using Microsoft.Extensions.Configuration;
-
-namespace Microsoft.BotBuilderSamples
-{
-    public class BotServices : IBotServices
-    {
-        public BotServices(IConfiguration configuration)
-        {
-            QnAMakerService = new QnAMaker(new QnAMakerEndpoint
-            {
-                KnowledgeBaseId = configuration["QnAKnowledgebaseId"],
-                EndpointKey = configuration["QnAEndpointKey"],
-                Host = GetHostname(configuration["QnAEndpointHostName"])
-            });
-        }
-
-        public QnAMaker QnAMakerService { get; private set; }
-
-        private static string GetHostname(string hostname)
-        {
-            if (!hostname.StartsWith("https://"))
-            {
-                hostname = string.Concat("https://", hostname);
-            }
-
-            if (!hostname.EndsWith("/qnamaker"))
-            {
-                hostname = string.Concat(hostname, "/qnamaker");
-            }
-
-            return hostname;
-        }
-    }
-}
