+++ samples/csharp_dotnetcore/11.qnamaker/Bots/QnABot.cs
@@ -1,52 +1,54 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-//
-// Generated with Bot Builder V4 SDK Template for Visual Studio EchoBot v4.5.0
 
-using System.Collections.Generic;
+using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Bot.Builder;
-using Microsoft.Bot.Builder.Dialogs;
+using Microsoft.Bot.Builder.AI.QnA;
 using Microsoft.Bot.Schema;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 
-namespace Microsoft.BotBuilderSamples.Bots
+namespace Microsoft.BotBuilderSamples
 {
-    public class QnABot<T> : ActivityHandler where T : Microsoft.Bot.Builder.Dialogs.Dialog
+    public class QnABot : ActivityHandler
     {
-        protected readonly BotState ConversationState;
-        protected readonly Microsoft.Bot.Builder.Dialogs.Dialog Dialog;
-        protected readonly BotState UserState;
+        private readonly IConfiguration _configuration;
+        private readonly ILogger<QnABot> _logger;
+        private readonly IHttpClientFactory _httpClientFactory;
 
-        public QnABot(ConversationState conversationState, UserState userState, T dialog)
+        public QnABot(IConfiguration configuration, ILogger<QnABot> logger, IHttpClientFactory httpClientFactory)
         {
-            ConversationState = conversationState;
-            UserState = userState;
-            Dialog = dialog;
+            _configuration = configuration;
+            _logger = logger;
+            _httpClientFactory = httpClientFactory;
         }
 
-        public override async Task OnTurnAsync(ITurnContext turnContext, CancellationToken cancellationToken = default)
+        protected override async Task OnMessageActivityAsync(ITurnContext<IMessageActivity> turnContext, CancellationToken cancellationToken)
         {
-            await base.OnTurnAsync(turnContext, cancellationToken);
+            var httpClient = _httpClientFactory.CreateClient();
 
-            // Save any state changes that might have occured during the turn.
-            await ConversationState.SaveChangesAsync(turnContext, false, cancellationToken);
-            await UserState.SaveChangesAsync(turnContext, false, cancellationToken);
-        }
-
-        protected override async Task OnMessageActivityAsync(ITurnContext<IMessageActivity> turnContext, CancellationToken cancellationToken) =>
-            // Run the Dialog with the new message Activity.
-            await Dialog.RunAsync(turnContext, ConversationState.CreateProperty<DialogState>(nameof(DialogState)), cancellationToken);
-
-        protected override async Task OnMembersAddedAsync(IList<ChannelAccount> membersAdded, ITurnContext<IConversationUpdateActivity> turnContext, CancellationToken cancellationToken)
-        {
-            foreach (var member in membersAdded)
+            var qnaMaker = new QnAMaker(new QnAMakerEndpoint
             {
-                if (member.Id != turnContext.Activity.Recipient.Id)
-                {
-                    await turnContext.SendActivityAsync(MessageFactory.Text($"Hello and welcome!"), cancellationToken);
-                }
+                KnowledgeBaseId = _configuration["QnAKnowledgebaseId"],
+                EndpointKey = _configuration["QnAEndpointKey"],
+                Host = _configuration["QnAEndpointHostName"]
+            },
+            null,
+            httpClient);
+
+            _logger.LogInformation("Calling QnA Maker");
+
+            // The actual call to the QnA Maker service.
+            var response = await qnaMaker.GetAnswersAsync(turnContext);
+            if (response != null && response.Length > 0)
+            {
+                await turnContext.SendActivityAsync(MessageFactory.Text(response[0].Answer), cancellationToken);
+            }
+            else
+            {
+                await turnContext.SendActivityAsync(MessageFactory.Text("No QnA Maker answers were found."), cancellationToken);
             }
         }
     }
