--- samples/csharp_dotnetcore/11.qnamaker/IBotServices.cs
@@ -1,12 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-using Microsoft.Bot.Builder.AI.QnA;
-
-namespace Microsoft.BotBuilderSamples
-{
-    public interface IBotServices
-    {
-        QnAMaker QnAMakerService { get; }
-    }
-}
