+++ samples/csharp_dotnetcore/11.qnamaker/Properties/launchSettings.json
