+++ samples/csharp_dotnetcore/11.qnamaker/appsettings.Development.json
@@ -1,9 +1,9 @@
 {
-    "Logging": {
-      "LogLevel": {
-        "Default": "Debug",
-        "System": "Information",
-        "Microsoft": "Information"
-      }
+  "Logging": {
+    "LogLevel": {
+      "Default": "Debug",
+      "System": "Information",
+      "Microsoft": "Information"
     }
   }
+}
