+++ samples/csharp_dotnetcore/11.qnamaker/appsettings.json
