+++ samples/csharp_dotnetcore/11.qnamaker/wwwroot/default.htm
@@ -4,7 +4,7 @@
 <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>QnAMakerBot</title>
+    <title>QnA Maker Sample</title>
     <style>
         body {
             margin: 0px;
@@ -88,20 +88,20 @@
             line-height: 32px;
         }
 
-        .main-text-p1 {
+        .main-text-p1{
             padding-top: 48px;
             padding-bottom: 28px;
         }
 
         .endpoint {
-            height: 32px;
-            width: 571px;
-            color: #808080;
-            font-family: "Segoe UI";
-            font-size: 24px;
-            font-weight: 200;
-            line-height: 32px;
-            padding-top: 28px;
+           height: 32px;
+           width: 571px;
+           color: #808080;
+           font-family: "Segoe UI";
+           font-size: 24px;
+           font-weight: 200;
+           line-height: 32px;
+           padding-top: 28px;
         }
 
         .how-to-build-section {
@@ -392,7 +392,7 @@
     <header class="header">
         <div class="header-inner-container">
             <div class="header-icon" style="display: inline-block"></div>
-            <div class="header-text" style="display: inline-block">QnAMakerBot Bot</div>
+            <div class="header-text" style="display: inline-block">QnA Maker Sample</div>
         </div>
     </header>
     <div class="row">
@@ -400,11 +400,9 @@
             <div class="content-title">Your bot is ready!</div>
             <div class="main-text main-text-p1">You can test your bot in the Bot Framework Emulator<br />
                 by connecting to http://localhost:3978/api/messages.</div>
-            <div class="main-text download-the-emulator"><a class="ctaLink"
-                    href="https://aka.ms/bot-framework-F5-download-emulator" target="_blank">Download the Emulator</a>
-            </div>
-            <div class="main-text">Visit <a class="ctaLink" href="https://aka.ms/bot-framework-F5-abs-home"
-                    target="_blank">Azure
+            <div class="main-text download-the-emulator"><a class="ctaLink" href="https://aka.ms/bot-framework-F5-download-emulator"
+                    target="_blank">Download the Emulator</a></div>
+            <div class="main-text">Visit <a class="ctaLink" href="https://aka.ms/bot-framework-F5-abs-home" target="_blank">Azure
                     Bot Service</a> to register your bot and add it to<br />
                 various channels. The bot's endpoint URL typically looks
                 like this:</div>
@@ -416,5 +414,4 @@
         <div class="ms-logo"></div>
     </div>
 </body>
-
 </html>
