--- samples/csharp_dotnetcore/11a.qnamaker/appsettings.json
@@ -1,7 +1,0 @@
-{
-  "MicrosoftAppId": "",
-  "MicrosoftAppPassword": "",
-  "QnAKnowledgebaseId": "",
-  "QnAEndpointKey": "",
-  "QnAEndpointHostName": ""
-}
