+++ samples/javascript_nodejs/11.qnamaker/.env
@@ -1,6 +1,5 @@
 MicrosoftAppId=
 MicrosoftAppPassword=
-
 QnAKnowledgebaseId=
 QnAEndpointKey=
 QnAEndpointHostName=
