+++ samples/javascript_nodejs/11.qnamaker/bots/QnABot.js
@@ -2,38 +2,21 @@
 // Licensed under the MIT License.
 
 const { ActivityHandler } = require('botbuilder');
+const { QnAMaker } = require('botbuilder-ai');
 
-/**
- * A simple bot that responds to utterances with answers from QnA Maker.
- * If an answer is not found for an utterance, the bot responds with help.
- */
 class QnABot extends ActivityHandler {
-    /**
-     *
-     * @param {ConversationState} conversationState
-     * @param {UserState} userState
-     * @param {Dialog} dialog
-     */
-    constructor(conversationState, userState, dialog) {
+    constructor() {
         super();
-        if (!conversationState) throw new Error('[QnABot]: Missing parameter. conversationState is required');
-        if (!userState) throw new Error('[QnABot]: Missing parameter. userState is required');
-        if (!dialog) throw new Error('[QnABot]: Missing parameter. dialog is required');
 
-        this.conversationState = conversationState;
-        this.userState = userState;
-        this.dialog = dialog;
-        this.dialogState = this.conversationState.createProperty('DialogState');
-
-        this.onMessage(async (context, next) => {
-            console.log('Running dialog with Message Activity.');
-
-            // Run the Dialog with the new message Activity.
-            await this.dialog.run(context, this.dialogState);
-
-            // By calling next() you ensure that the next BotHandler is run.
-            await next();
-        });
+        try {
+            this.qnaMaker = new QnAMaker({
+                knowledgeBaseId: process.env.QnAKnowledgebaseId,
+                endpointKey: process.env.QnAEndpointKey,
+                host: process.env.QnAEndpointHostName
+            });
+        } catch (err) {
+            console.warn(`QnAMaker Exception: ${ err } Check your QnAMaker configuration in .env`);
+        }
 
         // If a new user is added to the conversation, send them a greeting message
         this.onMembersAdded(async (context, next) => {
@@ -48,11 +31,31 @@ class QnABot extends ActivityHandler {
             await next();
         });
 
-        this.onDialog(async (context, next) => {
-            // Save any state changes. The load happened during the execution of the Dialog.
-            await this.conversationState.saveChanges(context, false);
-            await this.userState.saveChanges(context, false);
+        // When a user sends a message, perform a call to the QnA Maker service to retrieve matching Question and Answer pairs.
+        this.onMessage(async (context, next) => {
+            if (!process.env.QnAKnowledgebaseId || !process.env.QnAEndpointKey || !process.env.QnAEndpointHostName) {
+                let unconfiguredQnaMessage = 'NOTE: \r\n' + 
+                    'QnA Maker is not configured. To enable all capabilities, add `QnAKnowledgebaseId`, `QnAEndpointKey` and `QnAEndpointHostName` to the .env file. \r\n' +
+                    'You may visit www.qnamaker.ai to create a QnA Maker knowledge base.'
 
+                 await context.sendActivity(unconfiguredQnaMessage)
+            }
+            else {
+                console.log('Calling QnA Maker');
+    
+                const qnaResults = await this.qnaMaker.getAnswers(context);
+    
+                // If an answer was received from QnA Maker, send the answer back to the user.
+                if (qnaResults[0]) {
+                    await context.sendActivity(qnaResults[0].answer);
+    
+                // If no answers were returned from QnA Maker, reply with help.
+                } else {
+                    await context.sendActivity('No QnA Maker answers were found.');
+                }
+    
+            }
+            
             // By calling next() you ensure that the next BotHandler is run.
             await next();
         });
