--- samples/javascript_nodejs/11.qnamaker/dialogs/rootDialog.js
@@ -1,88 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-const {
-    ComponentDialog,
-    DialogSet,
-    DialogTurnStatus,
-    WaterfallDialog
-} = require('botbuilder-dialogs');
-
-const {
-    QnAMakerBaseDialog,
-    QNAMAKER_BASE_DIALOG,
-    DefaultCardNoMatchResponse,
-    DefaultCardNoMatchText,
-    DefaultCardTitle,
-    DefaultNoAnswer,
-    DefaultThreshold,
-    DefaultTopN,
-    QnAOptions,
-    QnADialogResponseOptions
-} = require('./qnamakerBaseDialog');
-
-const INITIAL_DIALOG = 'initial-dialog';
-const ROOT_DIALOG = 'root-dialog';
-
-class RootDialog extends ComponentDialog {
-    /**
-     * Root dialog for this bot.
-     * @param {QnAMaker} qnaService A QnAMaker service object.
-     */
-    constructor(qnaService) {
-        super(ROOT_DIALOG);
-
-        // Initial waterfall dialog.
-        this.addDialog(new WaterfallDialog(INITIAL_DIALOG, [
-            this.startInitialDialog.bind(this)
-        ]));
-
-        this.addDialog(new QnAMakerBaseDialog(qnaService));
-
-        this.initialDialogId = INITIAL_DIALOG;
-    }
-
-    /**
-     * The run method handles the incoming activity (in the form of a TurnContext) and passes it through the dialog system.
-     * If no dialog is active, it will start the default dialog.
-     * @param {*} turnContext
-     * @param {*} accessor
-     */
-    async run(context, accessor) {
-        const dialogSet = new DialogSet(accessor);
-        dialogSet.add(this);
-
-        const dialogContext = await dialogSet.createContext(context);
-        const results = await dialogContext.continueDialog();
-        if (results.status === DialogTurnStatus.empty) {
-            await dialogContext.beginDialog(this.id);
-        }
-    }
-
-    // This is the first step of the WaterfallDialog.
-    // It kicks off the dialog with the QnA Maker with provided options.
-    async startInitialDialog(step) {
-        // Set values for generate answer options.
-        var qnamakerOptions = {
-            scoreThreshold: DefaultThreshold,
-            top: DefaultTopN,
-            context: {}
-        };
-
-        // Set values for dialog responses.
-        var qnaDialogResponseOptions = {
-            noAnswer: DefaultNoAnswer,
-            activeLearningCardTitle: DefaultCardTitle,
-            cardNoMatchText: DefaultCardNoMatchText,
-            cardNoMatchResponse: DefaultCardNoMatchResponse
-        };
-
-        var dialogOptions = {};
-        dialogOptions[QnAOptions] = qnamakerOptions;
-        dialogOptions[QnADialogResponseOptions] = qnaDialogResponseOptions;
-
-        return await step.beginDialog(QNAMAKER_BASE_DIALOG, dialogOptions);
-    }
-}
-
-module.exports.RootDialog = RootDialog;
