+++ samples/javascript_nodejs/11.qnamaker/index.js
@@ -7,24 +7,17 @@
 const path = require('path');
 const restify = require('restify');
 
-// Import required bot services. See https://aka.ms/bot-services to learn more about the different parts of a bot.
-const { BotFrameworkAdapter, ConversationState, MemoryStorage, UserState } = require('botbuilder');
-const { QnAMaker } = require('botbuilder-ai');
+// Import required bot services.
+// See https://aka.ms/bot-services to learn more about the different parts of a bot.
+const { BotFrameworkAdapter } = require('botbuilder');
+const { ActivityTypes } = require('botbuilder-core');
 
+// The bot.
 const { QnABot } = require('./bots/QnABot');
-const { RootDialog } = require('./dialogs/rootDialog');
 
 // Note: Ensure you have a .env file and include QnAMakerKnowledgeBaseId, QnAMakerEndpointKey and QnAMakerHost.
 const ENV_FILE = path.join(__dirname, '.env');
 require('dotenv').config({ path: ENV_FILE });
-
-// Create HTTP server.
-const server = restify.createServer();
-server.listen(process.env.port || process.env.PORT || 3978, function() {
-    console.log(`\n${ server.name } listening to ${ server.url }.`);
-    console.log('\nGet Bot Framework Emulator: https://aka.ms/botframework-emulator');
-    console.log('\nTo talk to your bot, open the emulator select "Open Bot"');
-});
 
 // Create adapter.
 // See https://aka.ms/about-bot-adapter to learn more about adapters.
@@ -35,57 +28,43 @@ const adapter = new BotFrameworkAdapter({
 
 // Catch-all for errors.
 adapter.onTurnError = async (context, error) => {
+    // Create a trace activity that contains the error object
+    const traceActivity = {
+        type: ActivityTypes.Trace,
+        timestamp: new Date(),
+        name: 'onTurnError Trace',
+        label: 'TurnError',
+        value: `${ error }`,
+        valueType: 'https://www.botframework.com/schemas/error'
+    };
     // This check writes out errors to console log .vs. app insights.
     // NOTE: In production environment, you should consider logging this to Azure
     //       application insights.
     console.error(`\n [onTurnError] unhandled error: ${ error }`);
 
     // Send a trace activity, which will be displayed in Bot Framework Emulator
-    await context.sendTraceActivity(
-        'OnTurnError Trace',
-        `${ error }`,
-        'https://www.botframework.com/schemas/error',
-        'TurnError'
-    );
+    await context.sendActivity(traceActivity);
 
     // Send a message to the user
-    await context.sendActivity('The bot encounted an error or bug.');
-    await context.sendActivity('To continue to run this bot, please fix the bot source code.');
+    await context.sendActivity(`The bot encounted an error or bug.`);
+    await context.sendActivity(`To continue to run this bot, please fix the bot source code.`);
 };
 
-// Define the state store for your bot. See https://aka.ms/about-bot-state to learn more about using MemoryStorage.
-// A bot requires a state storage system to persist the dialog and user state between messages.
-const memoryStorage = new MemoryStorage();
+// Create the main dialog.
+const bot = new QnABot();
 
-// Create conversation and user state with in-memory storage provider.
-const conversationState = new ConversationState(memoryStorage);
-const userState = new UserState(memoryStorage);
-
-var endpointHostName = process.env.QnAEndpointHostName;
-if (!endpointHostName.startsWith('https://')) {
-    endpointHostName = 'https://' + endpointHostName;
-}
-
-if (!endpointHostName.endsWith('/qnamaker')) {
-    endpointHostName = endpointHostName + '/qnamaker';
-}
-
-const qnaService = new QnAMaker({
-    knowledgeBaseId: process.env.QnAKnowledgebaseId,
-    endpointKey: process.env.QnAEndpointKey,
-    host: endpointHostName
+// Create HTTP server
+const server = restify.createServer();
+server.listen(process.env.port || process.env.PORT || 3978, function() {
+    console.log(`\n${ server.name } listening to ${ server.url }`);
+    console.log(`\nGet Bot Framework Emulator: https://aka.ms/botframework-emulator`);
 });
 
-// Create the main dialog.
-const dialog = new RootDialog(qnaService);
-
-// Create the bot's main handler.
-const bot = new QnABot(conversationState, userState, dialog);
-
-// Listen for incoming requests.
+// Listen for incoming activities and route them to your bot main dialog.
 server.post('/api/messages', (req, res) => {
+    // Route received a request to adapter for processing
     adapter.processActivity(req, res, async (turnContext) => {
-        // Route the message to the bot's main handler.
+        // route to bot activity handler.
         await bot.run(turnContext);
     });
 });
