+++ samples/javascript_nodejs/11.qnamaker/package.json
@@ -1,14 +1,14 @@
 {
-    "name": "qnamaker-bot",
+    "name": "qnamaker",
     "version": "1.0.0",
-    "description": "Bot Builder v4 using QnA Maker service with active learning and multi-turn experience.",
+    "description": "Bot Builder v4 using QnA Maker service",
     "author": "Microsoft",
     "license": "MIT",
     "main": "index.js",
     "scripts": {
         "start": "node ./index.js",
         "watch": "nodemon ./index.js",
-        "lint": "./node_modules/.bin/eslint .",
+        "lint": "eslint .",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "repository": {
@@ -18,17 +18,16 @@
     "dependencies": {
         "botbuilder": "~4.6.0",
         "botbuilder-ai": "~4.6.0",
-        "botbuilder-dialogs": "~4.6.0",
-        "dotenv": "^8.2.0",
-        "restify": "^8.4.0"
+        "dotenv": "^8.0.0",
+        "restify": "~8.3.3"
     },
     "devDependencies": {
-        "eslint": "^6.6.0",
-        "eslint-config-standard": "^14.1.0",
+        "eslint": "^6.0.1",
+        "eslint-config-standard": "^13.0.1",
         "eslint-plugin-import": "^2.18.2",
-        "eslint-plugin-node": "^10.0.0",
+        "eslint-plugin-node": "^9.1.0",
         "eslint-plugin-promise": "^4.2.1",
-        "eslint-plugin-standard": "^4.0.1",
-        "nodemon": "~1.19.4"
+        "eslint-plugin-standard": "^4.0.0",
+        "nodemon": "~1.19.1"
     }
 }
