--- samples/javascript_nodejs/11a.qnamaker/.env
@@ -1,5 +1,0 @@
-MicrosoftAppId=
-MicrosoftAppPassword=
-QnAKnowledgebaseId=
-QnAEndpointKey=
-QnAEndpointHostName=
