--- samples/javascript_nodejs/11a.qnamaker/package.json
@@ -1,33 +1,0 @@
-{
-    "name": "qnamaker",
-    "version": "1.0.0",
-    "description": "Bot Builder v4 using QnA Maker service",
-    "author": "Microsoft",
-    "license": "MIT",
-    "main": "index.js",
-    "scripts": {
-        "start": "node ./index.js",
-        "watch": "nodemon ./index.js",
-        "lint": "eslint .",
-        "test": "echo \"Error: no test specified\" && exit 1"
-    },
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/Microsoft/BotBuilder-Samples.git"
-    },
-    "dependencies": {
-        "botbuilder": "~4.6.0",
-        "botbuilder-ai": "~4.6.0",
-        "dotenv": "^8.0.0",
-        "restify": "~8.3.3"
-    },
-    "devDependencies": {
-        "eslint": "^6.0.1",
-        "eslint-config-standard": "^13.0.1",
-        "eslint-plugin-import": "^2.18.2",
-        "eslint-plugin-node": "^9.1.0",
-        "eslint-plugin-promise": "^4.2.1",
-        "eslint-plugin-standard": "^4.0.0",
-        "nodemon": "~1.19.1"
-    }
-}
